pyproject.toml
[project]
name = "apigent"
version = "0.1.0"
description = "NL to API Agent"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "python-dotenv>=1.1.0",
    "jinja2>=3.1.6",
    "babel>=2.17.0",
    "psycopg[binary,pool]>=3.2.6",
    "pygments>=2.19.1",
    "pillow>=11.1.0",
    "openai>=1.72.0",
    "jieba>=0.42.1",
]
